+++ ChangeLog.md
@@ -1,8 +1,10 @@
 ## change log ##
 
 * 2016-06-21: 3.0.0-alpha:
-  - some code clean-up in main controller, now also reacting to evtAuthenticateStateChanged
-  - some code clean-up in user module, and new evtAuthenticateStateChanged
+  - some code clean-up and new evtAuthenticateStateChanged:
+    - admin-orgs directive
+    - orr-main directive
+    - orr-user directive
   - uiRoutes is only used for main page and associated actions (no need for rUri parameter)
 
 * 2016-06-20: 3.0.alpha: